Add WIP evidence transfers on login & helper sides
I'm sure there's a better way to send the files, but I will need to look
into it more. Also, evidence cannot yet be updated on the fly or any of
the fancy stuff I have planned yet, so there is still a ways to go
before this feature is complete.